Bundler fully implemented. Partials sample updated.
