+++ example/partial/_partial2.scss
@@ -1,1 +1,3 @@
+@import "non-existent";
+
 $partial-variable: red;
