+++ example/partial/main.scss
@@ -1,6 +1,5 @@
 @import "partial";
 
-
 .class {
     background-color: $partial-variable;
 }
