+++ example/simple/b.scss
@@ -1,1 +1,1 @@
-$b-variable: green;
+$b-variable: blue;
